% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_sim.R
\name{ci_sim}
\alias{ci_sim}
\title{ci_sim}
\usage{
ci_sim(n_cohorts, n_den, n_rev, n_case, N_sim)
}
\arguments{
\item{n_cohorts}{the number of cohorts, or buckets}

\item{n_den}{a vector containing denominator counts for each cohort}

\item{n_rev}{a vector containing the number of patients reviewed in each cohort}

\item{n_case}{a vector containing the number of hits in each cohort for the feasibility question of interest}

\item{N_sim}{the number of simulations}
}
\value{
a list containing a tibble with simulation output, a tibble with count percentiles, a tibble with summarized results intended for excel use, a summary tibble, and a histogram
}
\description{
This function performs bootstrap simulation for finding the confidence intervals of extrapolated feasibility counts with weighted proportions.
}
